moved to size_t for indices/sizes for larger arrays if needed added integer size to each data file added associated support in rawdataload allocation of output eigenstate memory is now done within the KineticOperator so that CrankNicolson can provide a lower memory solution. The user must now provide the solver with a pointer to the malloc'er and free the memory themselves movied file opening, writing, management for (most Measurers) to the Measurer base class. PotentialManager no longer allocates pointer arrays of zero length SimulationManager will use a default weight of 1 for each state if its weight calculator is not set instead of throwing an error added UniformWeight in case the user wants to control the value of the uniform weight moved Simulation stuff to submodule of tdsepy added Python bindings for DTBCs added general PySimulation.run function which determines the appropriate variant to use based upon the set KineticOperator and the types of Potentials
(known problems)
the Python wrapper structure is not fully functioning, hints are not being provided
program terminates with a segfault/double-free

+++ src/core/CuTridiagSolver.cpp
@@ -1,7 +1,7 @@
 #include "CuTridiagSolver.h"
 #include "cudaTools.cuh"
 
-cudaTridiagonalSolverSystem::cudaTridiagonalSolverSystem(int n, int nrhs) : n(n), nrhs(nrhs), lhsOffdiagDefined(false), rhsOffdiagDefined(false) {
+cudaTridiagonalSolverSystem::cudaTridiagonalSolverSystem(size_t n, size_t nrhs) : n(n), nrhs(nrhs), lhsOffdiagDefined(false), rhsOffdiagDefined(false) {
     cudaStatCheck(  cusparseCreate(&csHandle));
     cudaStatCheck(  cublasCreate_v2(&cbHandle));
     
@@ -135,13 +135,13 @@ void cudaTridiagonalSolverSystem::rhsProduct(const std::complex<double>* D, bool
     cuDoubleComplex zero = make_cuDoubleComplex(0.0, 0.0);
     // gbmv cannot work in-place, so if the source and dest are the same use a temporary buffer
     if(destVirt == sourceVirt){
-        for(int i = 0; i < nrhs; i++){
+        for(size_t i = 0; i < nrhs; i++){
             cudaStatCheck(  cublasZgbmv_v2(cbHandle, CUBLAS_OP_N, n, n, 1, 1, &one, cRHSMat, 3, sourceX.data+(i*n), 1, &zero, rhsTemp, 1));
             cudaStatCheck(  cublasZcopy_v2(cbHandle, n, rhsTemp, 1, destX.data+(i*n), 1));
         }
     }
     else{
-        for(int i = 0; i < nrhs; i++)
+        for(size_t i = 0; i < nrhs; i++)
             cudaStatCheck(  cublasZgbmv_v2(cbHandle, CUBLAS_OP_N, n, n, 1, 1, &one, cRHSMat, 3, sourceX.data+(i*n), 1, &zero, destX.data+(i*n), 1));
     }
